requirements.txt
# to simplify datetime parsing, used in task 1
python-dateutil